fix: harden runtime-view rehydration, replication framing, and gRPC plaintext channels (#1229)
* fix: harden runtime-view rehydration, replication framing, and gRPC plaintext channels

Three defence-in-depth fixes from a repository-wide security review:

- Runtime-view rehydration resolves persisted projection type names against
  an allow-list of loaded projection implementations instead of an
  unconstrained Type.GetType, closing an arbitrary-assembly-load primitive
  (#1226).
- The compressed-batch framing decoder widens its payload-bounds check to
  long arithmetic so a forged compressed-length cannot wrap past int.MaxValue
  and bypass the guard (#1227).
- The gRPC replication transports log a warning and increment a new
  orleans.lattice.replication.grpc.insecure_channel counter when
  AllowPlaintextEndpoints builds an insecure channel (#1228).

Each fix has regression tests; no public-API or wire-format change.

* feat: chart the gRPC insecure-channel metric on a bundled dashboard

Adds a Replication Transport (gRPC) Grafana dashboard
(LatticeDashboardKind.ReplicationGrpc) charting the new
orleans.lattice.replication.grpc.insecure_channel counter, and a
ReplicationGrpcMeterDashboardCoverageTests drift guard (deriving from the
shared MeterDashboardCoverageTestsBase) so every instrument on the gRPC
transport meter must be paneled - bringing the package in line with the
per-meter dashboard-coverage test pattern used by auth, membership,
scaling, and backup.

+++ src/lattice.replication/OrleansBinaryReplicationBatchEncoder.cs
@@ -565,7 +565,15 @@ internal sealed class OrleansBinaryReplicationBatchEncoder : IReplicationBatchEn
                 + "potential decompression bomb. Raise the ceiling if this reflects a legitimately large batch.",
                 nameof(payload));
         }
-        if (cursor + compressedLength > payload.Length)
+        // Widen to long before summing: both operands are int and
+        // compressedLength is an attacker-controllable wire field (up to
+        // int.MaxValue), so cursor + compressedLength can overflow to a
+        // negative value and slip past this truncation guard. The 64-bit
+        // sum cannot overflow for two non-negative int operands, so an
+        // oversized declared length is caught here by the explicit
+        // framing-corruption message rather than by the later
+        // span.Slice(cursor, compressedLength) bounds check.
+        if ((long)cursor + compressedLength > payload.Length)
         {
             throw new ArgumentException(
                 $"Framing payload is truncated at the compressed body; declared compressed length {compressedLength} would overrun the payload (remaining {payload.Length - cursor} bytes).",
